ffmpeg类型适配

+++ services/media_engine/plugins/ffmpeg_adapter/demuxer/ffmpeg_demuxer_plugin.h
@@ -82,7 +82,7 @@ private:
     Status ReadEosSample(std::shared_ptr<AVBuffer> sample);
     Status WriteBuffer(std::shared_ptr<AVBuffer> outBuffer, int64_t pts, uint32_t flag, const uint8_t *writeData,
         int32_t writeSize);
-    void ParseDrmInfo(const MetaDrmInfo *const metaDrmInfo, int32_t drmInfoSize,
+    void ParseDrmInfo(const MetaDrmInfo *const metaDrmInfo, size_t drmInfoSize,
         std::multimap<std::string, std::vector<uint8_t>>& drmInfo);
 
     struct IOContext {
